cleanup

+++ src/Coordinator.cc
@@ -71,7 +71,7 @@ void Coordinator::run() {
     steal_work();
     process_inbox();
 
-    if ((static_cast<int>(workers_idle.size()) == context.num_threads
+    if ((workers_idle.size() == static_cast<size_t>(context.num_threads)
           && context.assignments.size() == 0) || Coordinator::stopping) {
       stop_workers();
       // any worker that was running will have sent back a RETURN_WORK message
